Propagate dynamic memory address to read value.
This makes stack accesses unknown when SP is unknown.
It works as expected if SP is set to be known/static. However, this
introduces static addresses into captured code which only are valid
if the original memory region used as stack is still reserved.

We really would like to use the dynamically provided stack if possible,
and track static/dynamic property on this changing stack. This needs
a further tracking state...