Major refactoring of test script
The properties in the test case and the driver are now parsed properly
and multiple properties can be used at the same time. The script itself
is now object-oriented and has a proper argument parser.

If the success of a test is only defined by the return code, there is no
need to store an empty *.expect file for it. Consequently, a test case
is now defined via the *.c and the *.s files and not via the *.expect
files.

If the test case or the driver has the property nooutput=1, no output
verification or storing will be performed.