Fix Travis for decoder tests
The assembler of clang-3.5 in the images used by Travis seems to have
a bug with assembling the "test" instruction, resulting in machine
code with reversed operands (this does not happen e.g. with clang-3.8).
As decoder tests all use assembly files as input, we force use of the
GNU assembler "as" via setting the compile line accordingly for this
test class.