Add config to set parameter count
DBrew only supports up to 6 parameters (for now).
Now, one has to specify the par count for DBrew not to go wrong.

Update tests/test-driver.c : 2 int parameters for test function.
Also update examples/{stencil|simple|matrix}.c
matrix does not work (never did) as we only support 6 parameters.

TODO: add a mode which allows arbitrary number of parameters
without need to specify the count. This should work if we allow
the signature of the rewriter and the function to rewrite to be
exactly the same (no additional 1st par), and the emulation takes
parameters from stack if needed: with amd64 ABI, 7th par is on
stack...