Fix op-call test on Ubuntu 16.10
In Ubuntu 16.10, the linker option -pie (generate position-independent
code) is by default switched on. Before, function pointers in code
always were in the lower 4GB of a process, and 32bit was enough for
a relocation entry. With -pie however, this is no longer the case.

In the test case, the line "mov $jtarget, %r14" was a 32-bit move
and generated a R_X86_64_32S relocation. This fails with -pie now.
The test now uses "movabs", resulting in the use of a 64-bit immediate
operand and R_X86_64_64, which is fine with -pie (and of course also
works with -fno-pie (default in older distributions).

TODO:
It would be nice to enforce the linker options -pie/-no-pie, and we
could make 2 test cases: one with 32bit relocation, one with 64bit
relocation. However, gcc and clang work differently here, and we
need the test framework to pass different options depending on the
compiler (which is not supported for now): gcc forwards -pie/-no-pie
simply the linker, but also wants -fpie/-fno-pie as code generation
option (to only generate fitting relocations). clang instead DOES
NOT understand the linker options -pie/-no-pie, but instead adds them
for the linker automatically when -fpie/-fno-pie is given.