Improved and extended emulation of push/pop instructions
On the x86_64 architecture in 64-bit mode only 64- or 16-bit values can
interact with the stack (SDM vol. 2B, 4.3) since instructions commanding
otherwise are not encodable. Therefore, the previous support for
pushing/popping 32-bit registers (i.e. moving rsp by 4 bytes) was
removed. So now only 16-bit or 64-bit registers can be pushed from or
popped into. Furthermore, support for pushing and popping 8-bit and
16-bit immediate values has been added and the decoder now correctly
decodes push with a 16-bit immediate. 8-bit and 32-bit immediate values
are sign-extended to 64-bit. I'm not completely sure if this is the
right thing to do, but it appears to work in practice and gcc seems to
make the assumption that pushing any immediate value decrements rsp by 8.