Named memory/optional bytes for easier tests
When data is accessed by tests, the data addresses
used can change between compilers and compiler versions,
and even between multiple runs of the same test with
PIE (position independent executable) and address
randomization.

To make tests not to fail due to address changes between
a test run and expected output, the printer now may use
symbolic names and skip printing machine code bytes.

This is used in a test driver, and an corresponding test
is added.