Support for setting the known-ness of params. of arbitrary functions
Preliminary support for marking parameters of arbitrary functions static. The
added interface enables the known-ness of parameters of arbitrary functions to
be specified. The implementation builds on the memory range functionality and
generalizes the function configuration interface such that the handling of
rewriting entry functions and all other functions are unified.