SETcc: Preserve flags when zeroing register before SETcc
This is hack. Insert pushf/popf around the xor register zeroing instruction
added before a SETcc instruction in order to preserve the values of
flags. Replace this with a more general solution to handing registry zeroing
instructions (i.e. when is it safe to remove them and how to avoid removing them
when its not?