SETcc: Insert reg. load instr. before preceding cmp instr.
The SETcc instructions are frequently immediately preceded by a cmp
instruction. As the set instructions only set the last 8 its of the register it
operates on we need to load the value of the register if it is static. This
patch inserts the mov instruction to perform the load before the cmp instruction
is one is immediately preceding the set instruction thereby avoiding the use of
pushf/popf to save and restore the flags.