Make traversal order of paths stable
Processing of conditional jumps result in 2 follow-up paths
to travers when the condition is unknown. Previously, we used
the (unknown) condition value to decide in which order the
paths got traversed. This may make tests unreliable.

Also, traversal order relates to later BB generation order.
Now the order is deterministic and uses a static jump prediction
found in most processors: take backward but not forward branches.