cmp emu: Fix loading of static values to memory
For example, we have the instruction cmp %rax,0x20(%rbx) where the value of
0x20(%rbx) is 0x15 and static, %rbx is static and %rax is dynamic. Previously,
the following sequence was captured:

  mov 0x15,0x20(%rbx)
  cmp %rax,0x20(%rbx)

which is obviously wrong given that %rbx is static. This patch causes the
following code to be generated instead:

  mov <rbx static value>,%rbx
  cmp %rax,0x20(%rbx)

If dest indirect register is %rsp or %rbp, this transformation is not
performed (i.e. no value will be loaded into these registers)